major change - replace MDP algorithm Replaced MDP algorithm with a recursive version. NOTE: curveCollision() doesn't seem to be working, even with trivial       paths. Comment that function content to test MDP otherwise it       will fail every time.
MDP starts building the first and last dubins curve of the path and 
recursively go towards the middle of the path. When reaching a 
situation with one segment to build, the solution is trivial and 
positive if it does not collide. With 2 segments, the angle of the 
middle point is chosen between NUM_ANGLES dividing equally a circle
and using a simple heuristic. Shortest non colliding curves are chosen.